Make this work on Windows 11 & VS 2022

+++ src/instructions.cc
@@ -22,7 +22,7 @@
 // Helper functions
 
 // Set PC to @address if @condition is true
-void CPU::doBranch(bool condition, Word address, Byte &expectedCyclesToUse) {
+void CPU::doBranch(bool condition, Word address, uint64_t &expectedCyclesToUse) {
 	if (condition) {
 		Cycles++;	// Branch taken
 		expectedCyclesToUse++;
@@ -105,7 +105,7 @@ void CPU::doADC(Byte operand) {
 // CPU Instructions
 
 // ADC
-void CPU::ins_adc(Byte opcode, Byte &expectedCyclesToUse) {
+void CPU::ins_adc(Byte opcode, uint64_t &expectedCyclesToUse) {
 	Byte operand = getData(opcode, expectedCyclesToUse);
 	
 	if (Flags.D) {
@@ -117,7 +117,7 @@ void CPU::ins_adc(Byte opcode, Byte &expectedCyclesToUse) {
 }
 
 // AND
-void CPU::ins_and(Byte opcode, Byte &expectedCyclesToUse) {
+void CPU::ins_and(Byte opcode, uint64_t &expectedCyclesToUse) {
 	Byte data;
 
 	data = getData(opcode, expectedCyclesToUse);
@@ -127,7 +127,7 @@ void CPU::ins_and(Byte opcode, Byte &expectedCyclesToUse) {
 }
 
 // ASL
-void CPU::ins_asl(Byte opcode, Byte &expectedCyclesToUse) {
+void CPU::ins_asl(Byte opcode, uint64_t &expectedCyclesToUse) {
 	Word address;
 	Byte data;
 
@@ -155,7 +155,7 @@ void CPU::ins_asl(Byte opcode, Byte &expectedCyclesToUse) {
 }
 
 // BCC
-void CPU::ins_bcc(Byte opcode, Byte &expectedCyclesToUse) {
+void CPU::ins_bcc(Byte opcode, uint64_t &expectedCyclesToUse) {
 	Word address;
 
 	address = getAddress(opcode, expectedCyclesToUse);
@@ -163,7 +163,7 @@ void CPU::ins_bcc(Byte opcode, Byte &expectedCyclesToUse) {
 }
 
 // BCS
-void CPU::ins_bcs(Byte opcode, Byte &expectedCyclesToUse) {
+void CPU::ins_bcs(Byte opcode, uint64_t &expectedCyclesToUse) {
 	Word address;
 
 	address = getAddress(opcode, expectedCyclesToUse);
@@ -171,7 +171,7 @@ void CPU::ins_bcs(Byte opcode, Byte &expectedCyclesToUse) {
 }
 
 // BEQ
-void CPU::ins_beq(Byte opcode, Byte &expectedCyclesToUse) {
+void CPU::ins_beq(Byte opcode, uint64_t &expectedCyclesToUse) {
 	Word address;
 
 	address = getAddress(opcode, expectedCyclesToUse);
@@ -179,7 +179,7 @@ void CPU::ins_beq(Byte opcode, Byte &expectedCyclesToUse) {
 }
 
 // BIT
-void CPU::ins_bit(Byte opcode, Byte &expectedCyclesToUse) {
+void CPU::ins_bit(Byte opcode, uint64_t &expectedCyclesToUse) {
 	Byte data;
 
 	data = getData(opcode, expectedCyclesToUse);
@@ -189,7 +189,7 @@ void CPU::ins_bit(Byte opcode, Byte &expectedCyclesToUse) {
 }
 
 // BMI
-void CPU::ins_bmi(Byte opcode, Byte &expectedCyclesToUse) {
+void CPU::ins_bmi(Byte opcode, uint64_t &expectedCyclesToUse) {
 	Word address;
 
 	address = getAddress(opcode, expectedCyclesToUse);
@@ -197,7 +197,7 @@ void CPU::ins_bmi(Byte opcode, Byte &expectedCyclesToUse) {
 }
 
 // BNE
-void CPU::ins_bne(Byte opcode, Byte &expectedCyclesToUse) {
+void CPU::ins_bne(Byte opcode, uint64_t &expectedCyclesToUse) {
 	Word address;
 
 	address = getAddress(opcode, expectedCyclesToUse);
@@ -205,7 +205,7 @@ void CPU::ins_bne(Byte opcode, Byte &expectedCyclesToUse) {
 }
 
 // BPL
-void CPU::ins_bpl(Byte opcode, Byte &expectedCyclesToUse) {
+void CPU::ins_bpl(Byte opcode, uint64_t &expectedCyclesToUse) {
 	Word address;
 
 	address = getAddress(opcode, expectedCyclesToUse);
@@ -214,7 +214,7 @@ void CPU::ins_bpl(Byte opcode, Byte &expectedCyclesToUse) {
 
 // BRK
 void CPU::ins_brk([[maybe_unused]] Byte opcode,
-		  [[maybe_unused]] Byte &expectedCyclesToUse) {
+		  [[maybe_unused]] uint64_t &expectedCyclesToUse) {
 	// push PC + 1 to the stack. See:
 	// https://retrocomputing.stackexchange.com/questions/12291/what-are-uses-of-the-byte-after-brk-instruction-on-6502
 	addBacktrace(PC - 1);
@@ -225,7 +225,7 @@ void CPU::ins_brk([[maybe_unused]] Byte opcode,
 }
 
 // BVC
-void CPU::ins_bvc(Byte opcode, Byte &expectedCyclesToUse) {
+void CPU::ins_bvc(Byte opcode, uint64_t &expectedCyclesToUse) {
 	Word address;
 
 	address = getAddress(opcode, expectedCyclesToUse);
@@ -233,7 +233,7 @@ void CPU::ins_bvc(Byte opcode, Byte &expectedCyclesToUse) {
 }
 
 // BVS
-void CPU::ins_bvs(Byte opcode, Byte &expectedCyclesToUse) {
+void CPU::ins_bvs(Byte opcode, uint64_t &expectedCyclesToUse) {
 	Word address;
 
 	address = getAddress(opcode, expectedCyclesToUse);
@@ -242,34 +242,34 @@ void CPU::ins_bvs(Byte opcode, Byte &expectedCyclesToUse) {
 
 // CLC
 void CPU::ins_clc([[maybe_unused]] Byte opcode,
-		  [[maybe_unused]] Byte &expectedCyclesToUse) {
+		  [[maybe_unused]] uint64_t &expectedCyclesToUse) {
 	Flags.C = 0;
 	Cycles++;		// Single byte instruction
 }
 
 // CLD
 void CPU::ins_cld([[maybe_unused]] Byte opcode,
-		  [[maybe_unused]] Byte &expectedCyclesToUse) {
+		  [[maybe_unused]] uint64_t &expectedCyclesToUse) {
 	Flags.D = 0;
 	Cycles++;		// Single byte instruction
 }
 
 // CLI
 void CPU::ins_cli([[maybe_unused]] Byte opcode,
-		  [[maybe_unused]] Byte &expectedCyclesToUse) {
+		  [[maybe_unused]] uint64_t &expectedCyclesToUse) {
 	Flags.I = 0;
 	Cycles++;		// Single byte instruction
 }
 
 // CLV
 void CPU::ins_clv([[maybe_unused]] Byte opcode,
-		  [[maybe_unused]] Byte &expectedCyclesToUse) {
+		  [[maybe_unused]] uint64_t &expectedCyclesToUse) {
 	Flags.V = 0;
 	Cycles++;		// Single byte instruction
 }
 
 // CMP
-void CPU::ins_cmp(Byte opcode, Byte &expectedCyclesToUse) {
+void CPU::ins_cmp(Byte opcode, uint64_t &expectedCyclesToUse) {
 	Byte data;
 
 	data = getData(opcode, expectedCyclesToUse);
@@ -279,7 +279,7 @@ void CPU::ins_cmp(Byte opcode, Byte &expectedCyclesToUse) {
 }
 
 // CPX
-void CPU::ins_cpx(Byte opcode, Byte &expectedCyclesToUse) {
+void CPU::ins_cpx(Byte opcode, uint64_t &expectedCyclesToUse) {
 	Byte data;
 
 	data = getData(opcode, expectedCyclesToUse);
@@ -289,7 +289,7 @@ void CPU::ins_cpx(Byte opcode, Byte &expectedCyclesToUse) {
 }
 
 // CPY
-void CPU::ins_cpy(Byte opcode, Byte &expectedCyclesToUse) {
+void CPU::ins_cpy(Byte opcode, uint64_t &expectedCyclesToUse) {
 	Byte data;
 
 	data = getData(opcode, expectedCyclesToUse);
@@ -300,7 +300,7 @@ void CPU::ins_cpy(Byte opcode, Byte &expectedCyclesToUse) {
 }
 
 // DEC
-void CPU::ins_dec(Byte opcode, Byte &expectedCyclesToUse) {
+void CPU::ins_dec(Byte opcode, uint64_t &expectedCyclesToUse) {
 	Word address;
 	Byte data;
 
@@ -317,7 +317,7 @@ void CPU::ins_dec(Byte opcode, Byte &expectedCyclesToUse) {
 
 // DEX
 void CPU::ins_dex([[maybe_unused]] Byte opcode,
-		  [[maybe_unused]] Byte &expectedCyclesToUse) {
+		  [[maybe_unused]] uint64_t &expectedCyclesToUse) {
 	X--;
 	setFlagN(X);
 	setFlagZ(X);
@@ -326,7 +326,7 @@ void CPU::ins_dex([[maybe_unused]] Byte opcode,
 
 // DEY
 void CPU::ins_dey([[maybe_unused]] Byte opcode,
-		  [[maybe_unused]] Byte &expectedCyclesToUse) {
+		  [[maybe_unused]] uint64_t &expectedCyclesToUse) {
 	Y--;
 	setFlagN(Y);
 	setFlagZ(Y);
@@ -334,7 +334,7 @@ void CPU::ins_dey([[maybe_unused]] Byte opcode,
 }
 
 // EOR
-void CPU::ins_eor(Byte opcode, Byte &expectedCyclesToUse) {
+void CPU::ins_eor(Byte opcode, uint64_t &expectedCyclesToUse) {
 	Byte data;
 
 	data = getData(opcode, expectedCyclesToUse);
@@ -344,7 +344,7 @@ void CPU::ins_eor(Byte opcode, Byte &expectedCyclesToUse) {
 }
 
 // INC
-void CPU::ins_inc(Byte opcode, Byte &expectedCyclesToUse) {
+void CPU::ins_inc(Byte opcode, uint64_t &expectedCyclesToUse) {
 	Word address;
 	Byte data;
 
@@ -361,7 +361,7 @@ void CPU::ins_inc(Byte opcode, Byte &expectedCyclesToUse) {
 
 // INX
 void CPU::ins_inx([[maybe_unused]] Byte opcode,
-		  [[maybe_unused]] Byte &expectedCyclesToUse) {
+		  [[maybe_unused]] uint64_t &expectedCyclesToUse) {
 	X++;
 	setFlagZ(X);
 	setFlagN(X);
@@ -370,7 +370,7 @@ void CPU::ins_inx([[maybe_unused]] Byte opcode,
 
 // INY
 void CPU::ins_iny([[maybe_unused]] Byte opcode,
-		  [[maybe_unused]] Byte &expectedCyclesToUse) {
+		  [[maybe_unused]] uint64_t &expectedCyclesToUse) {
 	Y++;
 	setFlagZ(Y);
 	setFlagN(Y);
@@ -378,7 +378,7 @@ void CPU::ins_iny([[maybe_unused]] Byte opcode,
 }
 
 // JMP
-void CPU::ins_jmp(Byte opcode, Byte &expectedCyclesToUse) {
+void CPU::ins_jmp(Byte opcode, uint64_t &expectedCyclesToUse) {
 	Word address = PC;
 
 	addBacktrace(PC - 1);
@@ -392,7 +392,7 @@ void CPU::ins_jmp(Byte opcode, Byte &expectedCyclesToUse) {
 
 // JSR
 void CPU::ins_jsr([[maybe_unused]] Byte opcode,
-		  [[maybe_unused]] Byte &expectedCyclesToUse) {
+		  [[maybe_unused]] uint64_t &expectedCyclesToUse) {
 	Word newPC;
 	
 	addBacktrace(PC - 1);
@@ -405,28 +405,28 @@ void CPU::ins_jsr([[maybe_unused]] Byte opcode,
 }
 
 // LDA
-void CPU::ins_lda(Byte opcode, Byte &expectedCyclesToUse) {
+void CPU::ins_lda(Byte opcode, uint64_t &expectedCyclesToUse) {
 	A = getData(opcode, expectedCyclesToUse);
 	setFlagZ(A);
 	setFlagN(A);
 }
 
 // LDX
-void CPU::ins_ldx(Byte opcode, Byte &expectedCyclesToUse) {
+void CPU::ins_ldx(Byte opcode, uint64_t &expectedCyclesToUse) {
 	X = getData(opcode, expectedCyclesToUse);
 	setFlagZ(X);
 	setFlagN(X);
 }
 
 // LDY
-void CPU::ins_ldy(Byte opcode, Byte &expectedCyclesToUse) {
+void CPU::ins_ldy(Byte opcode, uint64_t &expectedCyclesToUse) {
 	Y = getData(opcode, expectedCyclesToUse);
 	setFlagZ(Y);
 	setFlagN(Y);
 }
 
 // LSR
-void CPU::ins_lsr(Byte opcode, Byte &expectedCyclesToUse) {
+void CPU::ins_lsr(Byte opcode, uint64_t &expectedCyclesToUse) {
 	Word address;
 	Byte data;
 
@@ -454,14 +454,14 @@ void CPU::ins_lsr(Byte opcode, Byte &expectedCyclesToUse) {
 
 // NOP
 void CPU::ins_nop([[maybe_unused]] Byte opcode,
-		  [[maybe_unused]] Byte &expectedCyclesToUse) {
+		  [[maybe_unused]] uint64_t &expectedCyclesToUse) {
 	// NOP, like all single byte instructions, takes
 	// two cycles.
 	Cycles++;
 }
 
 // ORA
-void CPU::ins_ora(Byte opcode, Byte &expectedCyclesToUse) {
+void CPU::ins_ora(Byte opcode, uint64_t &expectedCyclesToUse) {
 	Byte data;
 
 	data = getData(opcode, expectedCyclesToUse);
@@ -472,14 +472,14 @@ void CPU::ins_ora(Byte opcode, Byte &expectedCyclesToUse) {
 
 // PHA
 void CPU::ins_pha([[maybe_unused]] Byte opcode,
-		  [[maybe_unused]] Byte &expectedCyclesToUse) {
+		  [[maybe_unused]] uint64_t &expectedCyclesToUse) {
 	push(A);
 	Cycles++;		// Single byte instruction
 }
 
 // PLA
 void CPU::ins_pla([[maybe_unused]] Byte opcode,
-		  [[maybe_unused]] Byte &expectedCyclesToUse) {
+		  [[maybe_unused]] uint64_t &expectedCyclesToUse) {
 	A = pop();
 	setFlagN(A);
 	setFlagZ(A);
@@ -488,20 +488,20 @@ void CPU::ins_pla([[maybe_unused]] Byte opcode,
 
 // PHP
 void CPU::ins_php([[maybe_unused]] Byte opcode,
-		  [[maybe_unused]] Byte &expectedCyclesToUse) {
+		  [[maybe_unused]] uint64_t &expectedCyclesToUse) {
 	pushPS();
 	Cycles++;		// Single byte instruction
 }
 
 // PLP
 void CPU::ins_plp([[maybe_unused]] Byte opcode,
-		  [[maybe_unused]] Byte &expectedCyclesToUse) {
+		  [[maybe_unused]] uint64_t &expectedCyclesToUse) {
 	popPS();
 	Cycles += 2;
 }
 
 // ROL
-void CPU::ins_rol(Byte opcode, Byte &expectedCyclesToUse) {
+void CPU::ins_rol(Byte opcode, uint64_t &expectedCyclesToUse) {
 	Word address;
 	Byte data, carry;
 
@@ -531,7 +531,7 @@ void CPU::ins_rol(Byte opcode, Byte &expectedCyclesToUse) {
 }
 
 // ROR
-void CPU::ins_ror(Byte opcode, Byte &expectedCyclesToUse) {
+void CPU::ins_ror(Byte opcode, uint64_t &expectedCyclesToUse) {
 	Word address;
 	Byte data, zero;
 
@@ -562,7 +562,7 @@ void CPU::ins_ror(Byte opcode, Byte &expectedCyclesToUse) {
 
 // RTI
 void CPU::ins_rti([[maybe_unused]] Byte opcode,
-		  [[maybe_unused]] Byte &expectedCyclesToUse) {
+		  [[maybe_unused]] uint64_t &expectedCyclesToUse) {
 	removeBacktrace();
 	popPS();
 	PC = popWord();
@@ -571,7 +571,7 @@ void CPU::ins_rti([[maybe_unused]] Byte opcode,
 
 // RTS
 void CPU::ins_rts([[maybe_unused]] Byte opcode,
-		  [[maybe_unused]] Byte &expectedCyclesToUse) {
+		  [[maybe_unused]] uint64_t &expectedCyclesToUse) {
 	removeBacktrace();
 	
 	PC = popWord() + 1;
@@ -579,7 +579,7 @@ void CPU::ins_rts([[maybe_unused]] Byte opcode,
 }
 
 // SBC
-void CPU::ins_sbc(Byte opcode, Byte &expectedCyclesToUse) {
+void CPU::ins_sbc(Byte opcode, uint64_t &expectedCyclesToUse) {
 	Byte operand = getData(opcode, expectedCyclesToUse);
 
 	if (Flags.D) {
@@ -592,46 +592,46 @@ void CPU::ins_sbc(Byte opcode, Byte &expectedCyclesToUse) {
 
 // SEC
 void CPU::ins_sec([[maybe_unused]] Byte opcode,
-		  [[maybe_unused]] Byte &expectedCyclesToUse) {
+		  [[maybe_unused]] uint64_t &expectedCyclesToUse) {
 	Flags.C = 1;
 	Cycles++;		// Single byte instruction
 }
 
 // SED
 void CPU::ins_sed([[maybe_unused]] Byte opcode,
-		  [[maybe_unused]] Byte &expectedCyclesToUse) {
+		  [[maybe_unused]] uint64_t &expectedCyclesToUse) {
 	Flags.D = 1;
 	Cycles++;		// Single byte instruction
 }
 
 // SEI
 void CPU::ins_sei([[maybe_unused]] Byte opcode,
-		  [[maybe_unused]] Byte &expectedCyclesToUse) {
+		  [[maybe_unused]] uint64_t &expectedCyclesToUse) {
 	Flags.I = 1;
 	Cycles++;		// Single byte instruction
 }
 
 // STA
-void CPU::ins_sta(Byte opcode, Byte &expectedCyclesToUse) {
+void CPU::ins_sta(Byte opcode, uint64_t &expectedCyclesToUse) {
 	Word address = getAddress(opcode, expectedCyclesToUse);
 	writeByte(address, A);
 }
 
 // STX
-void CPU::ins_stx(Byte opcode, Byte &expectedCyclesToUse) {
+void CPU::ins_stx(Byte opcode, uint64_t &expectedCyclesToUse) {
 	Word address = getAddress(opcode, expectedCyclesToUse);
 	writeByte(address, X);
 }
 
 // STY
-void CPU::ins_sty(Byte opcode, Byte &expectedCyclesToUse) {
+void CPU::ins_sty(Byte opcode, uint64_t &expectedCyclesToUse) {
 	Word address = getAddress(opcode, expectedCyclesToUse);
 	writeByte(address, Y);
 }
 
 // TAX
 void CPU::ins_tax([[maybe_unused]] Byte opcode,
-		  [[maybe_unused]] Byte &expectedCyclesToUse) {
+		  [[maybe_unused]] uint64_t &expectedCyclesToUse) {
 	X = A;
 	setFlagZ(X);
 	setFlagN(X);
@@ -640,7 +640,7 @@ void CPU::ins_tax([[maybe_unused]] Byte opcode,
 
 // TAY
 void CPU::ins_tay([[maybe_unused]] Byte opcode,
-		  [[maybe_unused]] Byte &expectedCyclesToUse) {
+		  [[maybe_unused]] uint64_t &expectedCyclesToUse) {
 	Y = A;
 	setFlagZ(Y);
 	setFlagN(Y);
@@ -649,7 +649,7 @@ void CPU::ins_tay([[maybe_unused]] Byte opcode,
 
 // TSX
 void CPU::ins_tsx([[maybe_unused]] Byte opcode,
-		  [[maybe_unused]] Byte &expectedCyclesToUse) {
+		  [[maybe_unused]] uint64_t &expectedCyclesToUse) {
 	X = SP;
 	setFlagZ(X);
 	setFlagN(X);
@@ -658,7 +658,7 @@ void CPU::ins_tsx([[maybe_unused]] Byte opcode,
 
 // TXA
 void CPU::ins_txa([[maybe_unused]] Byte opcode,
-		  [[maybe_unused]] Byte &expectedCyclesToUse) {
+		  [[maybe_unused]] uint64_t &expectedCyclesToUse) {
 	A = X;
 	setFlagZ(A);
 	setFlagN(A);
@@ -667,14 +667,14 @@ void CPU::ins_txa([[maybe_unused]] Byte opcode,
 
 // TXS
 void CPU::ins_txs([[maybe_unused]] Byte opcode,
-		  [[maybe_unused]] Byte &expectedCyclesToUse) {
+		  [[maybe_unused]] uint64_t &expectedCyclesToUse) {
 	SP = X;
 	Cycles++;
 }
 
 // TYA
 void CPU::ins_tya([[maybe_unused]] Byte opcode,
-		  [[maybe_unused]] Byte &expectedCyclesToUse) {
+		  [[maybe_unused]] uint64_t &expectedCyclesToUse) {
 	A = Y;
 	setFlagZ(A);
 	setFlagN(A);
